Removing null QoS values from admin space entries.
It was not possible to prevent serde from serializing null values
with the #[serde(skip_serializing_if = "Option::is_none")] as
bincode is unable to serialize/deserialize such messages.

To avoid null QoS values being added to the admin space a filter
function was written to remove null values from the serde_json::Value
before returning the result of an admin space query.